Add more abstract interface to Docker
The commands are very tightly coupled to `pkg/docker` making them more
difficult to test. This is the beginning of hiding all the container
implementation details behind an interface, and then providing a Docker
implementation of that interface. This will allow the commands to be
tested with a mock implementation of the interface.