Fix bug causing nil-pointer dereferences
The way the global Viper config was being passed to the generator
function was causing a panic whenever `create` was run.

Part of this fix means not allowing an option for specifying the path to
the configuration file, which isn't a big deal. If it's something
necessary it can be added later.