Make variables easier to configure
The old way of configuring variables by making them raw strings with `=`
is cumbersome and counterintuitive. Really the only reason why it ever
was this way was because of a limitation in `viper` that Steve Francia
himself has said is expected behavior (see linked Github issue below).

In order to work around this, a custom configuration loader is
implemented. Another advantage to this is that the code is more
readable because now the YAML has an explicit schema and that schema is
backed by a type.

spf13/viper#260
UltimateSoftware/envctl#16