Show output of bootstrap on create
This is needed because users should be able to see what's going on as
their bootstrap steps are running. By definition, this also fixes the
bug where create returns before the bootstrap steps have actually
finished running.

This is done in much the same way that `login` attaches to the
container, but the terminal doesn't need to be put in raw mode and only
the hijacked reader is copied to stdout.

Resolves #11 and #10.