Add port bindings to environments
Sometimes users may want to run an app inside the environment, so they
can use their tooling to do so. In this case, they'll want to be able to
access the app on some exposed port.

This allows binding ports for different protocols by adding a key to the
config file that specifies them, grouped by protocol.